# backend/openapi.yaml
openapi: 3.1.0
info:
  title: StockSense AI Tools
  version: 1.0.0
  description: Tool endpoints for StockSense AI Agent Builder. Provides market data, portfolio context, watchlist state, and recommendation persistence for AI-driven investment signal generation.

servers:
  - url: http://localhost:3001
    description: Local development

paths:
  /tools/get_latest_price:
    get:
      operationId: getLatestPrice
      summary: Get the latest cached price for a ticker
      description: Returns the most recent price, volume, and change percentage for a single ticker from the latest_prices collection. Updated in real-time via Twelve Data WebSocket.
      parameters:
        - name: ticker
          in: query
          required: true
          description: Stock ticker symbol (e.g. AAPL, MSFT, NVDA)
          schema:
            type: string
            example: AAPL
      responses:
        "200":
          description: Latest price document for the ticker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatestPrice"
        "400":
          description: Missing ticker parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /tools/get_price_context:
    get:
      operationId: getPriceContext
      summary: Get technical price context for a ticker
      description: Computes 7-day and 30-day price change percentages, trend direction, 50/200-day moving average positioning, and volume spike detection using Twelve Data time_series API. Historical data is processed in memory and never stored.
      parameters:
        - name: ticker
          in: query
          required: true
          description: Stock ticker symbol
          schema:
            type: string
            example: AAPL
      responses:
        "200":
          description: Computed price context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceContext"
        "400":
          description: Missing ticker parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error or Twelve Data API failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /tools/get_market_news:
    get:
      operationId: getMarketNews
      summary: Get recent market news headlines for a ticker
      description: Returns up to 5 recent news headlines related to the specified ticker sourced from Yahoo Finance RSS. Used as context input for Gemini reasoning.
      parameters:
        - name: ticker
          in: query
          required: true
          description: Stock ticker symbol
          schema:
            type: string
            example: AAPL
      responses:
        "200":
          description: Array of news headlines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
        "400":
          description: Missing ticker parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /tools/get_portfolio:
    get:
      operationId: getPortfolio
      summary: Get all portfolio positions for a user
      description: Returns all holdings from the portfolio_positions collection for the specified user, including ticker, quantity, average purchase price, and sector.
      parameters:
        - name: userId
          in: query
          required: true
          description: User identifier
          schema:
            type: string
            example: verify-user
      responses:
        "200":
          description: Array of portfolio positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioPosition"
        "400":
          description: Missing userId parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /tools/get_watchlist:
    get:
      operationId: getWatchlist
      summary: Get all watchlist items for a user
      description: Returns all tickers being tracked across all of the user's watchlists from the watchlist_items collection. Includes ticker symbol and sector.
      parameters:
        - name: userId
          in: query
          required: true
          description: User identifier
          schema:
            type: string
            example: verify-user
      responses:
        "200":
          description: Array of watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
        "400":
          description: Missing userId parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /tools/save_recommendation:
    post:
      operationId: saveRecommendation
      summary: Save an AI-generated recommendation to the recommendation log
      description: Persists a structured recommendation (signal, confidence, rationale, factors, risks) to the recommendation_log collection. Used after the agent completes its reasoning pass for a ticker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRecommendationRequest"
      responses:
        "200":
          description: Saved recommendation document with generated _id and created_at
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error or validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /tools/record_feedback:
    post:
      operationId: recordFeedback
      summary: Record user feedback on a recommendation
      description: Updates a recommendation in the recommendation_log with the user's action (confirmed, ignored, or snoozed). This feedback drives the learning loop for future signal generation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordFeedbackRequest"
      responses:
        "200":
          description: Updated recommendation document with user_action and user_action_at
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error or validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required:
        - error

    LatestPrice:
      type: object
      properties:
        _id:
          type: string
        ticker:
          type: string
          example: AAPL
        price:
          type: number
          nullable: true
          example: 311.23
        volume:
          type: number
          nullable: true
          example: 37124759
        change_percent:
          type: number
          nullable: true
          example: 0.31
        updated_at:
          type: string
          format: date-time
          nullable: true
      required:
        - ticker

    PriceContext:
      type: object
      properties:
        current_price:
          type: number
          example: 311.23
        seven_day_change_pct:
          type: number
          nullable: true
          example: 0.941
        thirty_day_change_pct:
          type: number
          nullable: true
          example: 13.933
        trend:
          type: string
          enum: [bullish, bearish, neutral]
          example: neutral
        above_50dma:
          type: boolean
          nullable: true
          example: true
        above_200dma:
          type: boolean
          nullable: true
          example: true
        volume_spike:
          type: boolean
          example: false
      required:
        - current_price
        - trend
        - volume_spike

    NewsItem:
      type: object
      properties:
        headline:
          type: string
          example: "NVIDIA Could Seal a Major Alliance With Apple"
        published_at:
          type: string
          nullable: true
          example: "Thu, 04 Jun 2026 21:00:00 +0000"
      required:
        - headline

    PortfolioPosition:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        ticker:
          type: string
          example: AAPL
        quantity:
          type: number
          example: 10
        average_price:
          type: number
          example: 175.0
        sector:
          type: string
          nullable: true
          example: Technology
        current_value:
          type: number
          nullable: true
      required:
        - ticker
        - quantity
        - average_price

    WatchlistItem:
      type: object
      properties:
        _id:
          type: string
        watchlistId:
          type: string
        ticker:
          type: string
          example: AAPL
        sector:
          type: string
          nullable: true
          example: Technology
      required:
        - ticker

    SaveRecommendationRequest:
      type: object
      properties:
        userId:
          type: string
          description: User identifier
          example: verify-user
        ticker:
          type: string
          description: Stock ticker symbol
          example: AAPL
        signal:
          type: string
          enum: [BUY, HOLD, EXIT, WATCH, REBALANCE]
          description: Recommendation signal type
          example: BUY
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score between 0 and 1
          example: 0.85
        rationale:
          type: string
          description: Human-readable explanation of the recommendation
          example: Strong technical setup with positive news catalyst
        supporting_factors:
          type: array
          items:
            type: string
          description: Evidence supporting the recommendation
          example: ["Above 50DMA and 200DMA", "Positive earnings outlook"]
        risks:
          type: array
          items:
            type: string
          description: Identified risks to the recommendation
          example: ["Upcoming earnings volatility", "Sector rotation risk"]
      required:
        - userId
        - ticker
        - signal

    RecordFeedbackRequest:
      type: object
      properties:
        recId:
          type: string
          description: MongoDB _id of the recommendation to update
          example: 6a22ba9739e1bba9ea9fd11e
        user_action:
          type: string
          enum: [confirmed, ignored, snoozed]
          description: User's action on the recommendation
          example: confirmed
      required:
        - recId
        - user_action

    Recommendation:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        ticker:
          type: string
          example: AAPL
        signal:
          type: string
          enum: [BUY, HOLD, EXIT, WATCH, REBALANCE]
        confidence:
          type: number
          nullable: true
          example: 0.85
        rationale:
          type: string
        supporting_factors:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        user_action:
          type: string
          nullable: true
          enum: [confirmed, ignored, snoozed]
        user_action_at:
          type: string
          format: date-time
          nullable: true
      required:
        - userId
        - ticker
        - signal
        - created_at